code4ed-resume-analyser/requirements.txt
streamlit>=1.28.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=0.8.11
docx2txt>=0.8
spacy>=3.7.0
nltk>=3.8.1
scikit-learn>=1.3.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
chromadb>=0.4.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.20.0
numpy>=1.24.0
pandas>=2.1.0
plotly>=5.0.0
